# Optimized brigand::remove and brigand::remove_if