```cpp
using list = brigand::remove<brigand::range<int, 0, SIZE>, std::integral_constant<int, SIZE/2>>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 36M | 00.09s - 38M | 00.12s - 41M | 00.20s - 46M | 00.33s - 58M
gcc | 00.05s - 21M | 00.09s - 26M | 00.13s - 33M | 00.22s - 46M | 00.40s - 73M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 36M | 00.07s - 36M | 00.07s - 37M | 00.09s - 38M | 00.12s - 42M
gcc | 00.04s - 20M | 00.05s - 20M | 00.06s - 21M | 00.07s - 23M | 00.11s - 28M

---

```cpp
using list = brigand::remove_if<brigand::range<int, 0, SIZE>, is_odd<brigand::_1>>;
```
Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.09s - 38M | 00.13s - 41M | 00.20s - 46M | 00.35s - 57M
gcc | 00.05s - 21M | 00.09s - 26M | 00.13s - 32M | 00.22s - 44M | 00.39s - 68M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 36M | 00.09s - 38M | 00.13s - 40M | 00.19s - 45M | 00.34s - 55M
gcc | 00.05s - 21M | 00.08s - 25M | 00.12s - 30M | 00.20s - 42M | 00.37s - 64M