@ldionne 2016-03-24
@jonathanpoelen On a side note, how do you measure the RAM usage of the compiler? I need it for [Metabench](https://github.com/ldionne/metabench).