@edouarda 2016-03-24
@ldionne on Linux you can read /proc