@ldionne 2016-03-24
Thanks. I wish there was a cross-platform way to do that.