@edouarda 2016-03-24
Windows API is pretty easy to use. https://msdn.microsoft.com/en-us/library/windows/desktop/ms686832(v=vs.85).aspx