@jonathanpoelen 2016-03-25
I use `/usr/bin/time --format="%E %M"` on Linux, easier than read /proc.